+++ rust/src/structs/mod.rs
@@ -1,3 +1,4 @@
+pub mod message;
 pub mod messages;
 pub mod header;
 pub mod command;